Remove `start_url` from manifest.json
This allows clients to use whatever url the user is currently at when i.e. adding the PWA to their home screen. Together with the functionality of opening a connection via a url string (i.e. `client.local/#192.168.1.103:10000`) opening the app shortcut would send the user straight to the entity that they want to control, avoiding them needing to remember and input an IP address/hostname first.